(almost) Finished direction changes in walk cycle
The only issue is that I didn't consider the fact that we need a separate sprite for whenever moving upwards, since the perspective is different.  This would only work if the POV was completely top-down.  I need to fix this, and also make the sprite return to a default image whenever not moving.