Added draw_bound
This will be a dev-only program that will pikl an array of rectangles that will serve as the boundary lines in the world.  This is going to be a seperate program so that there's no way for users to interfere with the boundaries and break the game.  Might be deleted, might not.  Only intended for developers regardless