Fixing walk cycle in multiple gamestates
Current bug where holding down a direction and then pausing before letting go results in the speed being added upon, and never returning back to 0.  Need to change logic within the entity class so that movement speed is set equal to a seperate parameter, and x and y dictate what direction the sprite is moving.